fix(climate): Optimize ecoregion envelope extraction using JSON method
- Replace slow CROSS JOIN with get_climate_json_at_point() function
- Fix ROUND type casting for PERCENTILE_CONT in analysis script
- Reduces extraction time from 40+ minutes to ~30 seconds
- Successfully populated 46,767 ecoregion envelopes (81.7% tree coverage)

+++ scripts/analyze-envelope-discrepancies.sql
@@ -254,8 +254,8 @@ LIMIT 20;
 \echo '=== GBIF vs TreeGOER Comparison (Trees with Both Sources) ==='
 SELECT
     COUNT(*) as species_with_both,
-    ROUND(AVG(ABS(ceg.temp_mean - cee.temp_mean)), 2) as avg_temp_diff,
-    ROUND(PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY ABS(ceg.temp_mean - cee.temp_mean)), 2) as median_temp_diff,
+    ROUND(AVG(ABS(ceg.temp_mean - cee.temp_mean))::numeric, 2) as avg_temp_diff,
+    ROUND(PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY ABS(ceg.temp_mean - cee.temp_mean))::numeric, 2) as median_temp_diff,
     COUNT(*) FILTER (WHERE ABS(ceg.temp_mean - cee.temp_mean) <= 2) as within_2c,
     COUNT(*) FILTER (WHERE ABS(ceg.temp_mean - cee.temp_mean) <= 5) as within_5c,
     COUNT(*) FILTER (WHERE ABS(ceg.temp_mean - cee.temp_mean) > 5) as beyond_5c
